+++ src/hooks/useRoleCapabilities/index.js
@@ -1,0 +1,1 @@
+export { useRoleCapabilities } from './useRoleCapabilities';