+++ src/hooks/useRoleCapabilitySets/index.js
@@ -1,0 +1,1 @@
+export { useRoleCapabilitySets } from './useRoleCapabilitySets';